Type the TTL as int32, which is what it has always been
gosec G115 on both call sites: integer overflow conversion int -> int32.

The conversion existed because the config field was declared int while its
only consumer is batchv1.JobSpec.TTLSecondsAfterFinished, which is *int32.
Declaring it int32 removes two casts and the warning together -- the field
was never an int in any meaningful sense.

Also checks the error from os.Unsetenv in the new config test. The merged
config excludes errcheck in _test.go, so CI would not have flagged it, but
an unchecked Unsetenv in a test that depends on the variable being absent
is worth checking on its own account.

Linted the way CI does before pushing -- base config yq-merged with the
repo's own, not the repo's alone: 0 issues. Running the repo config by
itself is STRICTER than CI and reported three pre-existing errcheck hits in
internal/logging/logging_test.go that the merged config excludes. Chasing
those would have been work against a rule this repo does not run.

+++ cmd/server/main.go
@@ -119,7 +119,7 @@ func run() error {
 			GCSKeySecret:               cfg.DispatchGCSKeySecret,
 			ClusterID:                  cfg.ClusterID,
 			ActiveDeadlineSeconds:      int64(cfg.DispatchTimeout().Seconds()),
-			JobTTLSecondsAfterFinished: int32(cfg.JobTTLSecondsAfterFinished),
+			JobTTLSecondsAfterFinished: cfg.JobTTLSecondsAfterFinished,
 			RepoHost:                   cfg.DispatchRepoHost,
 			RepoOrg:                    cfg.DispatchRepoOrg,
 			RefFallbackTemplates:       strings.Split(cfg.DispatchRefFallbacksRaw, "|"),
@@ -143,7 +143,7 @@ func run() error {
 	var forensicsDispatcher *forensics.Dispatcher
 	if cfg.ForensicsEnabled && cfg.ForensicsRunnerImage != "" {
 		forensicsDispatcher = forensics.New(forensics.Config{
-			JobTTLSecondsAfterFinished: int32(cfg.JobTTLSecondsAfterFinished),
+			JobTTLSecondsAfterFinished: cfg.JobTTLSecondsAfterFinished,
 			Enabled:                    cfg.ForensicsEnabled,
 			RunnerImage:                cfg.ForensicsRunnerImage,
 			TempoBaseURL:               cfg.ForensicsTempoBaseURL,
